4_Django_Forum_Project/requirements.txt
asgiref==3.7.2
Django==4.2.7
djangorestframework==3.14.0
Pygments==2.16.1
pytz==2023.3.post1
sqlparse==0.4.4
tzdata==2023.3
